Refactor Refresh Data to extract all common functions into a parent class

+++ database/migrations/2016_01_19_080502_create_episodes_table.php
@@ -21,7 +21,7 @@ class CreateEpisodesTable extends Migration
             $table->string('status');
             $table->string('airdate');
             $table->text('description');
-            $table->integer('file_size');
+            $table->bigInteger('file_size');
             $table->string('location');
             $table->timestamps();
         });
